Use better values for apt options.
apt considers DIRECT as a special value for Proxy. The only reason none
is not used is because the URI parser will not set the host part of it.

Just as well, Proxy-Auto-Detect value is the command to run to detect a
proxy. So, it should really be set to the empty value. Older apt
versions required using an extra empty argument to set empty values. But
newer versions interpret such an extra argument as the action, so it
doesn't work. See bug #693092 for details. However, the documentation
claims and the code shows that if there is a proxy set for the given
host, including the DIRECT option, Proxy-Auto-Detect is ignored and not
run at all.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@debian.org>