Check for explicitly configured proxies before trying the gateway
In some networks the hosts might have severely limited networking
access to the exterior, including to their gateway host. If the local
network administration had made configuration to provide a proxy
explicitly, honor that.

This may cause delays in other use cases where DNS and/or mDNS are slow
to respond, but that will need to be worked around in some other way, by
either being more aggressive without timeouts.

Closes: #1108265