Check explicitly configured proxies before network gateway
In some networks the hosts might have severely limited networking
access to the exterior, including to their gateway host. If the local
network administration had made configuration to provide a proxy
explicitly, honor that, otherwise users can see severe delays in APT
usage, or have it not work at all.

This is minimal backport for trixie of the original change (c2d0159:
"Check for explicitly configured proxies before trying the gateway").
That fix is applied after quite some refactoring, so it's not directly
applicable to the trixie version. Instead, we duplicate a bit of code to
achieve the same effect.

Closes: #1108265